rohit code

+++ src/main/java/com/hcl/payee/entity/Payee.java
@@ -9,19 +9,19 @@ import javax.persistence.Id;
 public class Payee {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
-	
+	private long id;
+
 	private String name;
 	private String branch;
 	private long accountNumber;
 	private String accountType;
 	private long otp;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
